Eliminate api/query -> api -> ... -> api/checks dep chain
searchcache service build is failing because it erroneously depends on api/checks and its template files. This change factors a helper function out to the shared package to break the dep chain.
