+++ api/results_redirect_handler.go
@@ -7,8 +7,6 @@ package api
 import (
 	"fmt"
 	"net/http"
-	"regexp"
-	"strings"
 
 	"github.com/web-platform-tests/wpt.fyi/shared"
 )
@@ -44,21 +42,7 @@ func apiResultsRedirectHandler(w http.ResponseWriter, r *http.Request) {
 	}
 
 	test := r.URL.Query().Get("test")
-	resultsURL := GetResultsURL(allRuns[0], test)
+	resultsURL := shared.GetResultsURL(allRuns[0], test)
 
 	http.Redirect(w, r, resultsURL, http.StatusFound)
-}
-
-// GetResultsURL constructs the URL to the result of a single test file in the given run.
-func GetResultsURL(run shared.TestRun, testFile string) (resultsURL string) {
-	resultsURL = run.ResultsURL
-	if testFile != "" && testFile != "/" {
-		// Assumes that result files are under a directory named SHA[0:10].
-		resultsBase := strings.SplitAfter(resultsURL, "/"+run.Revision)[0]
-		resultsPieces := strings.Split(resultsURL, "/")
-		re := regexp.MustCompile("(-summary)?\\.json\\.gz$")
-		product := re.ReplaceAllString(resultsPieces[len(resultsPieces)-1], "")
-		resultsURL = fmt.Sprintf("%s/%s/%s", resultsBase, product, testFile)
-	}
-	return resultsURL
 }
